Finally got the project working.
