--- src/main/java/com/reefat/gamesite/HelloWorld.java
@@ -1,31 +1,0 @@
-package com.reefat.gamesite;
-
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-
-/**
- * A simple REST service which is able to say hello to someone using HelloService Look at the web.xml where JAX-RS is enabled
- */
-
-@Path("/")
-public class HelloWorld {
-    @Inject
-    HelloService helloService;
-
-    @GET
-    @Path("/json")
-    @Produces({ "application/json" })
-    public String getHelloWorldJSON() {
-        return "{\"result\":\"" + helloService.createHelloMessage("World") + "\"}";
-    }
-
-    @GET
-    @Path("/xml")
-    @Produces({ "application/xml" })
-    public String getHelloWorldXML() {
-        return "<xml><result>" + helloService.createHelloMessage("World") + "</result></xml>";
-    }
-
-}
