+++ src/main/webapp/index.html
@@ -1,24 +1,7 @@
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>gamesite</title>
+<title>helloworld-rs</title>
 </head>
 <body>
 
@@ -26,7 +9,7 @@
 
     <ul>
     	<li><a href="rest/json">JSON</a></li>
-        <li style="font-size: 50px; color: red;"> Amit m8 where are you</li>
+        <li style="color: blue;"> AMITTO SAN!</li>
     </ul>
 </body>
 </html>
